Added "color" property to Sprite
If set during instantiation, Sprite will now draw a rectangle defined by
its rect() of color. (Because it is using fillStyle, this can also be a
linear or radial gradient too.)

Width and height, which now default to null, must also be set during an
instantiation of Sprite when the property color is also set for anything
to appear. (Will fail silently.)