Fixed Sprite, Text Testing; Added QuadTree tests
For some reason, jaws.Sprite has lost its "color" option and the
adjusted canvas drawing. This is fixing it again. (It's probably my
fault, but it's still weird it seems to keep changing.)

Re-added text.js (so its tests will now pass) and added quadtree.js so
it can be tested.

Added tests for QuadTree.