Fixed text bug (removed translate)
Finally fixed a bug in Text I noticed a few days ago by removing the
initial translate before the drawing of the text.

Because test is drawn using a x and y, the additional translate (unlike
sprites) frequently moves the final drawing position to incorrect
values. This is much more noticeable when drawing texts within
viewports.