Quadtree.js: Fixes, Docs
An update to quadtree.js that switches the call to
jaws.collideOneWithMany with the simplified jaws.collide with a
callback. It also now supports any object that has a forEach (both
SpriteList and Array-like objects). There is also testing in the more
public insert and retrieve functions.

There is also an expansion of all documentation from simple comments to
full JSDocs format for each function.