Fix for Safari 5.1.7
There is bug in the current version of Safari that will return a slightly off result from the currentTime property of the HTML5 Audio object.

Seeking the file to fraction will result in a value that is the fraction plus the audio buffer offset. (A difference of 0.00004 in most cases.)

This passes the test for Safari.