Remove SpriteList, fix for collide(list, list)
This removes the SpriteList instance and makes it just an array. This also adds another array, sprites2, to use for the collide(list, list) testing.

It turns out that in some browsers (notably IE and Safari), the testing would register twice. First, with the right values and then again with the false ones. Firefox and Chrome would take the first result and IE and Safari (and possibly others) would take the second.